Implemented utility method for frequency options and improved function name and comments (#32j3gu0)
